Create utils file to contain facility routes utility functions
